Frame Rate Display
When I program games in Pygame, I often want an easy tool to measure
frame rate, so I made my own for entertainment and educational purposes.
The class is deliberately simple.  You declare an instance of the class
and pass it a location to display at and a rate at which you want it to
update, and it detects and displays the frame rate appropriately.